Update styles.css
Step 8 Passed

Time for CSS.

Even though your <div> has no text, it's still treated as a box with content. Write a CSS rule that uses the .canvas class selector and set its width to 500 pixels. Here's a CSS rule that sets the width of the class card to 300 pixels:

.card {
  width: 300px;
}